fix(colors): slightly lighten up the darkest `--contrast-` swatch
Previously, the `--kompendium-contrast-1700` in the `light` color-scheme,
and `--kompendium-contrast-100` in the `dark` color-scheme were rendered
fully black, equivalent of `rgb(0,0,0)`.
We had to change make it lighter to improve the color palette, because:
1. We already have a swatch for fully black, with a CSS variable
called `--kompendium-color-black`.
2. Having these darkest `--contrast-` swatches render fully black
makes it impossible to
visualize shadows or other black elements on it. Since this swatch is
intended to be used for UI elements, it is important that it
provides a level of safety as well, so in case any fully black element
such as a piece of text that is hardcoded to be black is placed on it,
the user can still have a chance to see it.