fix(playground): use the darkest contrast swatch for the bg of examples in `dark` mode
Normally on  a user interface, in the `light` color-scheme,
most components will be placed on top of the
lightest or one of the lightest contrast swatches.
This means when we switch to the `dark` color-scheme
the components will be on the darkest or one of the darkest
contrast swatches.
Which is why it is better not to intentionally change
the background of the `show-case_component` area.
This is where users of Kompendium will be testing their
components and examples in the generated documentations.